Break out `featureTreeMachine`
